+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/schemas/type_header_shape.rs
