+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/shape_request_bodies.rs
@@ -1,4 +1,4 @@
-use crate::conversions::v3_0::to_rust_type::components::from_request_bodies::{
+use crate::conversions::v3_0::to_rust_type::components::request_bodies::{
     ContentShape, DefinitionShape, MediaTypeShape,
 };
 use crate::conversions::v3_0::to_rust_type::components::{create_module, ComponentsShapes};
